Separate components into individual files
